cmd/livefuzzer: split blob and basic fuzzing commands
